Various design improvements
