+++ RadialGauge/samples/RadialGaugeTemplatedStyleCustomSample.xaml
@@ -18,9 +18,9 @@
 
     <StackPanel Spacing="16">
         <controls:RadialGauge_ClassicBinding FontFamily="{x:Bind TextFontFamily, Mode=OneWay}"
-                                                 FontSize="{x:Bind TextSize, Mode=OneWay}"
-                                                 Foreground="{x:Bind TextForeground, Mode=OneWay}"
-                                                 Style="{StaticResource CustomClassicBindingRadialGaugeStyle}"
-                                                 Visibility="{x:Bind IsTextVisible, Mode=OneWay}" />
+                                             FontSize="{x:Bind TextSize, Mode=OneWay}"
+                                             Foreground="{x:Bind TextForeground, Mode=OneWay}"
+                                             Style="{StaticResource CustomClassicBindingRadialGaugeStyle}"
+                                             Visibility="{x:Bind IsTextVisible, Mode=OneWay}" />
     </StackPanel>
 </Page>
