+++ RadialGauge/samples/RadialGaugeXbindBackedStyleCustomSample.xaml
@@ -18,9 +18,9 @@
 
     <StackPanel Spacing="16">
         <controls:RadialGauge_xBind FontFamily="{x:Bind TextFontFamily, Mode=OneWay}"
-                                        FontSize="{x:Bind TextSize, Mode=OneWay}"
-                                        Foreground="{x:Bind TextForeground, Mode=OneWay}"
-                                        Style="{StaticResource CustomxBindRadialGaugeStyle}"
-                                        Visibility="{x:Bind IsTextVisible, Mode=OneWay}" />
+                                    FontSize="{x:Bind TextSize, Mode=OneWay}"
+                                    Foreground="{x:Bind TextForeground, Mode=OneWay}"
+                                    Style="{StaticResource CustomxBindRadialGaugeStyle}"
+                                    Visibility="{x:Bind IsTextVisible, Mode=OneWay}" />
     </StackPanel>
 </Page>
