+++ components/RadialGauge/samples/RadialGaugeCustomSample.xaml
@@ -1,4 +1,4 @@
-<!--  Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information.  -->
+﻿<!--  Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information.  -->
 <Page x:Class="RadialGaugeExperiment.Samples.RadialGaugeCustomSample"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -8,28 +8,31 @@
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
       mc:Ignorable="d">
 
-    <StackPanel Spacing="16">
-        <TextBlock Style="{StaticResource BodyStrongTextBlockStyle}" />
-        <StackPanel Spacing="16">
-            <controls:RadialGauge x:Name="RadialGauge"
-                                  Grid.Column="1"
-                                  IsInteractive="True"
-                                  MaxAngle="150"
-                                  Maximum="240"
-                                  MinAngle="210"
-                                  Minimum="0"
-                                  NeedleLength="100"
-                                  NeedleWidth="4"
-                                  ScalePadding="23"
-                                  ScaleTickWidth="4"
-                                  ScaleWidth="4"
-                                  StepSize="13"
-                                  TickLength="10"
-                                  TickSpacing="30"
-                                  TickWidth="4"
-                                  Unit="units"
-                                  ValueStringFormat="N0"
-                                  Value="94" />
-        </StackPanel>
+    <StackPanel HorizontalAlignment="Center"
+                VerticalAlignment="Center"
+                Orientation="Horizontal">
+        <controls:RadialGauge x:Name="RadialGauge"
+                              Width="280"
+                              IsEnabled="{x:Bind Enabled, Mode=OneWay}"
+                              IsInteractive="{x:Bind IsInteractive, Mode=OneWay}"
+                              MaxAngle="{x:Bind (x:Int32)MaxAngle, Mode=OneWay}"
+                              Maximum="240"
+                              MinAngle="{x:Bind (x:Int32)MinAngle, Mode=OneWay}"
+                              Minimum="0"
+                              NeedleLength="{x:Bind NeedleLength, Mode=OneWay}"
+                              NeedleWidth="{x:Bind NeedleWidth, Mode=OneWay}"
+                              ScalePadding="{x:Bind ScalePadding, Mode=OneWay}"
+                              ScaleTickWidth="{x:Bind ScaleTickWidth, Mode=OneWay}"
+                              ScaleWidth="{x:Bind ScaleWidth, Mode=OneWay}"
+                              StepSize="{x:Bind StepSize, Mode=OneWay}"
+                              TickLength="{x:Bind TickLength, Mode=OneWay}"
+                              TickPadding="{x:Bind TickPadding, Mode=OneWay}"
+                              TickSpacing="{x:Bind (x:Int32)TickSpacing, Mode=OneWay}"
+                              TickWidth="{x:Bind TickWidth, Mode=OneWay}"
+                              ValueStringFormat="N0"
+                              Value="{x:Bind Value, Mode=OneWay}" />
+        <controls:RadialGauge Width="280"
+                              Maximum="240"
+                              Minimum="0" />
     </StackPanel>
 </Page>
